fix: Divide FontAwesome & Searchbar
