+++ src/components/atoms/FontAwesome/index.ts
@@ -1,0 +1,3 @@
+import FontAwesome from "./FontAwesome";
+
+export default FontAwesome;
