+++ src/components/molecules/index.ts
@@ -1,3 +1,4 @@
 import HeaderRight from "./HeaderRight";
+import SearchBar from "./SearchBar";
 
-export { HeaderRight };
+export { HeaderRight, SearchBar };
